requirements/dev.txt
-i https://pypi.org/simple
-r base.txt
black==24.4.2
django-stubs[compatible-mypy]==5.0.2
factory-boy==3.3.0
flake8==7.1.0
ipython==8.25.0
isort==5.13.2
mypy==1.10.0
pylint-django==2.5.5
pylint==3.2.3
pytest-django==4.8.0
pytest==8.2.2
pytest-dotenv==0.5.2
raincoat==1.2.4
types-Markdown==3.6.0.20240316
types-requests<2.31.0.7 # See note in https://pypi.org/project/types-requests/ as opensearch-py 2.5.0 requires urllib3<2,>=1.26.18